Update telegram_channel_scraper.py
Added special char support for keywords. 
A line was added to escape special characters in each keyword using re.escape(). This function ensures that special characters in the keywords are treated as literal characters when used in regex searches. This is important because characters like ., *, ?, (, ), [, ], {, }, ^, $, |, and \ have special meanings in regex and can lead to unexpected behaviour if not properly escaped.